Switch Host::send to Region

+++ backend/include/WorkerInfo.h
@@ -16,7 +16,7 @@ struct WorkerInfo {
     /**
      * The time that computation has taken in nanoseconds
      */
-    long long computationTime;
+    long computationTime;
 
     /**
      * The region that was computed.
